Robustness update
Interview pointed out a bunch of areas where it would be extremely easy to improve the robustness of the code.  I sorta faceplanted on those questions but have since implemented consts where possible and adding a bunch of data validity/array boundary checks to handle errors in the input file.